+++ CHANGELOG.md
@@ -4,13 +4,47 @@ This changelog follows [keep a changelog](https://keepachangelog.com/) guideline
 
 The date format is `mm/dd/yyyy`.
 
-# 10/6/2023 - 1.0.9
+# 10/6/2023 - 1.1.0
+
+## social-sharing
+
+### ⬆️ Improved
+
+- Now uses `ShareCompat.IntentBuilder` instead of manually creating the intent (Learn more about [ShareCompat.IntentBuilder](https://developer.android.com/reference/androidx/core/app/ShareCompat.IntentBuilder))
+- Added `@Keep` to the `UnityFileProvider` class to prevent it from being removed by optimizations
+
+### ⚠️ Changed
+
+- Renamed `setUri()` to `setStream()`
+- Renamed `addUri()` to `addStream()`
+- Renamed `setUri()` to `setStream()`
+- Renamed `setSharesheetTitle()` to `setChooserTitle()`
+
+### ❌ Removed
+
+- Removed `useSingle()`
+- Removed `useMultiple()`
+- Removed `setThumbnailUri()`
+- Removed `setThumbnailUriFromString()`
+- Removed `setUriFromString()`
+- Removed `addUriFromString()`
+- Removed `setUseRichPreview()`
+- Removed `setUseSharesheet()`
 
 ## social-sharing-unity
 
 ### 🐞 Fixed
 
 - Replaced the call to the `FileProvider` in the `getUriFromPathUnity()` method of `SocialSharingUnity` class with `UnityFileProvider`
+
+### ⚠️ Changed
+
+- Renamed `setUnityUri()` to `setUnityStream()`
+- Renamed `addUnityUri()` to `addUnityStream()`
+
+### ❌ Removed
+
+- Removed `setUnityThumbnailUri()`
 
 # 10/5/2023 - 1.0.6